refactor: split chat command families
